refactor: 기본 생성자 접근 제한자 protected로 변경 및 패키지 이동
