+++ src/main/java/com/prgrms/springJpa/domain/customer/Customer.java
@@ -1,9 +1,10 @@
-package com.prgrms.springJpa.domain;
+package com.prgrms.springJpa.domain.customer;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
+import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,7 +14,7 @@ import lombok.Setter;
 @Table
 @Getter
 @Setter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class Customer {
